chore(connect): circular dep cleanup
