+++ packages/connect/src/data/defaultFeeLevels.ts
@@ -1,6 +1,6 @@
 import { BigNumber, typedObjectKeys } from '@trezor/utils';
 
-import { FeeInfo, FeeLevel } from '../types';
+import type { FeeInfo, FeeLevel } from '../types/fees';
 
 // this is workaround for the lack of information from 'trezor-common'
 // we need to declare what does "high/normal/low" mean in block time (eg: normal BTC = 6 blocks = ~1 hour)
