+++ packages/connect/src/types/settings.ts
@@ -4,7 +4,7 @@ import type { ThpCredentials, ThpPairingMethod } from '@trezor/protocol';
 import type { Transport } from '@trezor/transport';
 import { PartialRecord } from '@trezor/type-utils';
 
-import { FirmwareUpdateSource } from '../data/firmwareInfo';
+import type { FirmwareUpdateSource } from '../types/firmware';
 
 export type { SystemInfo } from '@trezor/connect-common';
 export interface Manifest {
