+++ packages/connect/src/utils/firmwareReleaseConfigUtils.ts
@@ -4,7 +4,8 @@ import { FirmwareReleaseConfig } from '@trezor/device-utils';
 import { getFirmwareReleaseJwsPublicKey } from '@trezor/env-utils';
 
 import { firmwareReleaseConfigAssets } from './assetUtils';
-import { FirmwareUpdateSource, getOnlineFirmwareBaseUrl } from '../data/firmwareInfo';
+import { getOnlineFirmwareBaseUrl } from '../data/firmwareInfo';
+import { FirmwareUpdateSource } from '../types/firmware';
 
 const JWS_CONFIG = {
     SIGN_ALGORITHM: 'ES256',
